+++ packages/create-app-container/src/createTabBar.ts
@@ -79,6 +79,7 @@ const createTabBar = (api) => (props) => {
     history.listen((location) => {
       setPathname(location.pathname);
     });
+  // eslint-disable-next-line
   }, []);
 
   return createElement(
@@ -103,7 +104,9 @@ const createTabBar = (api) => (props) => {
               key: `tab-${index}`,
               style: styles.tabBarItem,
               onClick: () => {
-                onClick && onClick(item);
+                if (onClick) {
+                  onClick(item);
+                }
                 if (!item.path) {
                   console.warn(`TabBar item ${item.name} need set path`);
                 } else {