+++ packages/plugin-app-base/src/userConfig/extensions.js
@@ -1,0 +1,8 @@
+module.exports = (config, extensions) => {
+  if (Array.isArray(extensions)) {
+    extensions.forEach((extension) => {
+      config.resolve.extensions
+        .add(extension);
+    });
+  }
+};