+++ packages/plugin-miniapp/src/index.ts
@@ -1,24 +1,24 @@
 import * as miniappConfig from 'miniapp-runtime-config';
-import * as builderShared from 'miniapp-builder-shared';
 import * as path from 'path';
 
 module.exports = (api) => {
-  const { onGetWebpackConfig, context, getValue } = api;
+  const { onGetWebpackConfig, context, registerUserConfig } = api;
   const { rootDir, userConfig } = context;
   const { targets = [] } = userConfig;
 
   targets.forEach((target) => {
     if (target === 'miniapp' || target === 'wechat-miniprogram') {
       onGetWebpackConfig(target, (config) => {
-        const projectType = getValue('PROJECT_TYPE');
-        // Clear entry
-        config.entryPoints.clear();
-        // App entry
-        config.entry('index').add(builderShared.pathHelper.getDepPath(rootDir, `app.${projectType}`));
+        const { outputDir = 'build' } = userConfig;
+        registerUserConfig({
+          name: target,
+          validation: 'object'
+        });
 
-        config.output.path(path.join(rootDir, 'build'));
+        const outputPath = path.resolve(rootDir, outputDir, target);
+        config.output.path(path.join(rootDir, 'build', target));
 
-        miniappConfig.setConfig(config, userConfig[target] || {}, { context, target, babelRuleName: 'babel-loader', onGetWebpackConfig });
+        miniappConfig.setConfig(config, userConfig[target] || {}, { context, target, babelRuleName: 'babel-loader', outputPath });
 
         if (config.plugins.get('MiniCssExtractPlugin')) {
           config.plugin('MiniCssExtractPlugin').tap((args) => [